Most of the Cornerstone now compiles after the QString change

+++ multitude/Luminous/RenderTarget.cpp
@@ -129,7 +129,7 @@ namespace Luminous
   std::shared_ptr<RenderTargetManager::RenderTargetState> RenderTargetManager::allocateNewTexture(size_t extent)
   {
     Radiant::info("RenderTargetManager::allocateNewTexture # %ld next %ld",
-                  extent, (int) (1 << nextHigherPowerOfTwo(extent)));
+                  extent, (long) (1 << nextHigherPowerOfTwo(extent)));
 
     // Use power-of-two textures
     extent = 1 << nextHigherPowerOfTwo(extent);
